fix: remove stale /ai-monitor references across backend, LLM, tests, and docs (#1004)
- Update LLM prompt context and navigate_to tool to use /health instead of /ai-monitor
- Update activity feed insight links to /health
- Replace /ai-monitor with /health in LANDING_PAGE_OPTIONS
- Fix stale test fixtures (settings, users, activity-feed, llm-tools)
- Update README.md and docs/architecture.md page names
- Add E2E test for /ai-monitor → /health redirect
- Extract FleetHealthSummary into own component (ai-monitor.tsx 1275→1114 lines)
