+++ packages/ai-intelligence/src/services/llm-tools.ts
@@ -195,7 +195,7 @@ export const TOOL_DEFINITIONS: ToolDefinition[] = [
             'stacks',
             'images',
             'networks',
-            'ai-monitor',
+            'health',
             'remediation',
             'logs',
             'metrics',
@@ -686,7 +686,7 @@ function executeNavigateTo(
     stacks: '/stacks',
     images: '/images',
     networks: '/networks',
-    'ai-monitor': '/ai-monitor',
+    health: '/health',
     remediation: '/remediation',
     logs: '/logs',
     metrics: '/metrics',
